[Feat] #4 - Moya Base 파일들 수정
